requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
parameterized==0.9.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
